+++ src/main/java/com/msg/after_school/domain/after_school/data/dto/ApplyAfterSchoolDto.java
@@ -1,0 +1,12 @@
+package com.msg.after_school.domain.after_school.data.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+@AllArgsConstructor
+public class ApplyAfterSchoolDto {
+    private final Long afterSchoolId;
+}
