+++ src/main/java/com/msg/after_school/domain/after_school/data/entity/ClassRegistration.java
@@ -1,0 +1,23 @@
+package com.msg.after_school.domain.after_school.data.entity;
+
+import com.msg.after_school.domain.user.entity.User;
+import lombok.Getter;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+public class ClassRegistration {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "afterschool_id")
+    private AfterSchool afterSchool;
+
+    @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "users_email")
+    private User user;
+}
+
