+++ src/main/java/com/msg/after_school/domain/after_school/data/type/SeasonType.java
@@ -1,0 +1,10 @@
+package com.msg.after_school.domain.after_school.data.type;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+
+@Getter
+@AllArgsConstructor
+public enum SeasonType {
+    FIRST,SECOND,SUMMER,WINTER
+}
