+++ src/main/java/com/msg/after_school/domain/after_school/response/AfterSchoolResponse.java
@@ -1,0 +1,8 @@
+package com.msg.after_school.domain.after_school.response;
+
+import lombok.Builder;
+
+@Builder
+public class AfterSchoolResponse {
+
+}
