+++ src/main/java/com/msg/after_school/domain/after_school/util/AfterSchoolConverter.java
@@ -1,0 +1,10 @@
+package com.msg.after_school.domain.after_school.util;
+
+import com.msg.after_school.domain.after_school.data.dto.AfterSchoolDto;
+import com.msg.after_school.domain.after_school.response.AfterSchoolResponse;
+
+import java.util.List;
+
+public interface AfterSchoolConverter {
+    List<AfterSchoolResponse> toResponse(List<AfterSchoolDto> dtoList);
+}
