+++ src/main/java/com/msg/after_school/domain/user/exception/UserNotFoundException.java
@@ -1,10 +1,10 @@
 package com.msg.after_school.domain.user.exception;
 
-import com.msg.after_school.global.error.exception.AfterSchoolException;
+import com.msg.after_school.global.error.exception.GlobalException;
 import com.msg.after_school.global.error.exception.ErrorCode;
 
-public class UserNotFoundException extends AfterSchoolException {
-    public static final AfterSchoolException EXCEPTION = new UserNotFoundException();
+public class UserNotFoundException extends GlobalException {
+    public static final GlobalException EXCEPTION = new UserNotFoundException();
 
     private UserNotFoundException() {
         super(ErrorCode.USER_NOT_FOUND);
