+++ src/main/java/com/msg/after_school/global/error/exception/GlobalException.java
@@ -1,0 +1,10 @@
+package com.msg.after_school.global.error.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class GlobalException extends RuntimeException{
+    private final ErrorCode errorCode;
+}
