+++ src/main/java/com/msg/after_school/global/security/auth/AuthDetailsService.java
@@ -3,7 +3,6 @@ package com.msg.after_school.global.security.auth;
 import com.msg.after_school.domain.user.exception.UserNotFoundException;
 import com.msg.after_school.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
