+++ src/main/java/com/msg/after_school/global/security/exception/ExpiredTokenException.java
@@ -1,10 +1,10 @@
 package com.msg.after_school.global.security.exception;
 
-import com.msg.after_school.global.error.exception.AfterSchoolException;
+import com.msg.after_school.global.error.exception.GlobalException;
 import com.msg.after_school.global.error.exception.ErrorCode;
 
-public class ExpiredTokenException extends AfterSchoolException {
-    public static final AfterSchoolException EXCEPTION = new ExpiredTokenException();
+public class ExpiredTokenException extends GlobalException {
+    public static final GlobalException EXCEPTION = new ExpiredTokenException();
 
     private ExpiredTokenException() {
         super(ErrorCode.EXPIRED_TOKEN);
