+++ src/main/java/com/msg/after_school/global/security/exception/handler/JwtExceptionHandler.java
@@ -2,10 +2,8 @@ package com.msg.after_school.global.security.exception.handler;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.msg.after_school.global.error.ErrorResponse;
-import com.msg.after_school.global.error.exception.AfterSchoolException;
+import com.msg.after_school.global.error.exception.GlobalException;
 import com.msg.after_school.global.error.exception.ErrorCode;
-import com.msg.after_school.global.security.exception.ExpiredTokenException;
-import com.msg.after_school.global.security.exception.InvalidTokenException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
@@ -27,7 +25,7 @@ public class JwtExceptionHandler extends OncePerRequestFilter {
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
         try{
             filterChain.doFilter(request,response);
-        }catch(AfterSchoolException e) {
+        }catch(GlobalException e) {
             log.error(request.getRequestURI());
             log.error(e.getErrorCode().getMessage());
             e.printStackTrace();
