+++ src/main/java/com/msg/after_school/global/security/filter/JwtTokenFilter.java
@@ -1,9 +1,7 @@
 package com.msg.after_school.global.security.filter;
 
-import com.msg.after_school.global.error.exception.AfterSchoolException;
 import com.msg.after_school.global.security.JwtTokenProvider;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.filter.OncePerRequestFilter;
