+++ src/main/java/com/msg/after_school/global/user/facade/UserFacade.java
@@ -1,0 +1,23 @@
+package com.msg.after_school.global.user.facade;
+
+import com.msg.after_school.domain.user.entity.User;
+import com.msg.after_school.domain.user.exception.UserNotFoundException;
+import com.msg.after_school.domain.user.repository.UserRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.stereotype.Component;
+
+@Component
+@RequiredArgsConstructor
+public class UserFacade {
+    private final UserRepository userRepository;
+
+    public User getCurrentUser() {
+        String email = SecurityContextHolder.getContext().getAuthentication().getName();
+        return getUserByEmail(email);
+    }
+
+    public User getUserByEmail(String email) {
+        return userRepository.findUserByEmail(email).orElseThrow(() -> UserNotFoundException.EXCEPTION);
+    }
+}
