--- src/main/java/com/msg/after_school/repository/DayOfWeekRepository.java
@@ -1,13 +1,0 @@
-package com.msg.after_school.repository;
-
-import com.msg.after_school.domain.DayOfWeek;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface DayOfWeekRepository extends JpaRepository<DayOfWeek,Long> {
-    List<DayOfWeek> findDayOfWeekByDayOfWeek(String dayOfWeek);
-
-}
