--- src/main/java/com/msg/after_school/response/AfterSchoolResponse.java
@@ -1,7 +1,0 @@
-package com.msg.after_school.response;
-
-import lombok.Builder;
-
-@Builder
-public class AfterSchoolResponse {
-}
