--- src/main/java/com/msg/after_school/service/AfterSchoolService.java
@@ -1,10 +1,0 @@
-package com.msg.after_school.service;
-
-import com.msg.after_school.dto.AfterSchoolDto;
-import com.msg.after_school.dto.SearchConditionDto;
-
-import java.util.List;
-
-public interface AfterSchoolService {
-    List<AfterSchoolDto> findAfterSchoolListBySearchCondition(SearchConditionDto dto);
-}
