--- src/main/java/com/msg/after_school/type/SeasonType.java
@@ -1,10 +1,0 @@
-package com.msg.after_school.type;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum SeasonType {
-    FIRST,SECOND,SUMMER,WINTER
-}
